# [PERF] Cache results of State.get