--- a/lib/route-recognizer.js
+++ b/lib/route-recognizer.js
@@ -159,10 +159,15 @@ function parse(route, names, specificity) {
 function State(charSpec) {
   this.charSpec = charSpec;
   this.nextStates = [];
+  this.charSpecs = {};
 }
 
 State.prototype = {
   get: function(charSpec) {
+    if (this.charSpecs[charSpec.validChars]) {
+      return this.charSpecs[charSpec.validChars];
+    }
+
     var nextStates = this.nextStates;
 
     for (var i=0; i<nextStates.length; i++) {
@@ -171,7 +176,10 @@ State.prototype = {
       var isEqual = child.charSpec.validChars === charSpec.validChars;
       isEqual = isEqual && child.charSpec.invalidChars === charSpec.invalidChars;
 
-      if (isEqual) { return child; }
+      if (isEqual) {
+        this.charSpecs[charSpec.validChars] = child;
+        return child;
+      }
     }
   },