Speed up of about ~30 op/s consistently overall. 5000 op/s improvement around `recognize`. Tried using an object that was placed into dictionary mode for the cache as the shape changes overtime but seemed slower.
```
  Add ................... 505.78 op/s
  End-to-end ............ 270.97 op/s
  Generate ........ 4,874,011.82 op/s
  Handlers For ... 37,588,869.18 op/s
  Recognize ......... 360,278.12 op/s
```